requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-imagehorizonlibrary
opencv-python
Pillow